# Significant `TokenMap` optimisation by removing clones and returning refs instead.