--- a/sway-lsp/src/capabilities/code_actions/diagnostic/auto_import.rs
+++ b/sway-lsp/src/capabilities/code_actions/diagnostic/auto_import.rs
@@ -41,15 +41,14 @@ pub(crate) fn import_code_action(
     let mut program_type_keyword = None;
 
     ctx.tokens.tokens_for_file(ctx.temp_uri).for_each(|item| {
-        // TODO remove this clone
-        let token = item.value().clone();
-        if let Some(TypedAstToken::TypedUseStatement(use_stmt)) = token.typed {
-            use_statements.push(use_stmt);
-        } else if let Some(TypedAstToken::TypedIncludeStatement(include_stmt)) = token.typed {
-            include_statements.push(include_stmt);
-        } else if token.kind == SymbolKind::ProgramTypeKeyword {
-            if let AstToken::Keyword(ident) = token.parsed {
-                program_type_keyword = Some(ident);
+        if let Some(TypedAstToken::TypedUseStatement(use_stmt)) = &item.value().typed {
+            use_statements.push(use_stmt.clone());
+        } else if let Some(TypedAstToken::TypedIncludeStatement(include_stmt)) = &item.value().typed
+        {
+            include_statements.push(include_stmt.clone());
+        } else if item.value().kind == SymbolKind::ProgramTypeKeyword {
+            if let AstToken::Keyword(ident) = &item.value().parsed {
+                program_type_keyword = Some(ident.clone());
             }
         }
     });