--- a/sway-lsp/src/capabilities/document_symbol.rs
+++ b/sway-lsp/src/capabilities/document_symbol.rs
@@ -1,18 +1,21 @@
 use crate::core::token::{SymbolKind, Token, TokenIdent};
+use dashmap::mapref::multiple::RefMulti;
 use lsp_types::{self, Location, SymbolInformation, Url};
 
-pub fn to_symbol_information<I>(tokens: I, url: Url) -> Vec<SymbolInformation>
+
+pub fn to_symbol_information<'a, I>(
+    tokens: I,
+    url: Url,
+) -> Vec<SymbolInformation>
 where
-    I: Iterator<Item = (TokenIdent, Token)>,
+    I: Iterator<Item = RefMulti<'a, TokenIdent, Token>>,
 {
-    let mut symbols: Vec<SymbolInformation> = vec![];
-
-    for (ident, token) in tokens {
-        let symbol = symbol_info(&ident, &token, url.clone());
-        symbols.push(symbol)
-    }
-
-    symbols
+    tokens
+        .map(|entry| {
+            let (ident, token) = entry.pair();
+            symbol_info(ident, token, url.clone()) // Assuming symbol_info can take &TokenIdent and &Token
+        })
+        .collect()
 }
 
 /// Given a `token::SymbolKind`, return the `lsp_types::SymbolKind` that corresponds to it.