## Description
This PR refactors the functions implemented on `TokenMap` to remove clones where possible, and instead return the `RefMulti` type from `DashMap` directly. As can be seen below, we are getting significant performance wins and most LSP capabilities are now measured in µs. 🚀 

related to #5445

| Metric                    | Before          | After         | Change (%) |
|---------------------------|-----------------|---------------|------------|
| tokens_for_file           | 9.8117 ms       | 253.845 µs    | -97.411%   |
| idents_at_position        | 3.6331 ms       | 47.870 µs     | -98.692%   |
| tokens_at_position        | 6.9532 ms       | 1.1478 ms     | -83.474%   |
| token_at_position         | 5.9395 ms       | 208.23 µs     | -96.483%   |
| parent_decl_at_position   | 6.9780 ms       | 1.1633 ms     | -83.428%   |
| semantic_tokens           | 18.2205 ms      | 587.64 µs     | -96.767%   |
| document_symbol           | 6.6865 ms       | 584.33 µs     | -91.186%   |
| completion                | 14.0025 ms      | 1.9808 ms     | -85.855%   |
| hover                     | 6.2409 ms       | 317.90 µs     | -94.902%   |
| highlight                 | 14.8390 ms      | 2.7572 ms     | -81.417%   |
| goto_definition           | 5.9403 ms       | 225.65 µs     | -96.204%   |
| inlay_hints               | 5.9934 ms       | 318.41 µs     | -94.677%   |
| prepare_rename            | 5.8056 ms       | 207.62 µs     | -96.411%   |
| rename                    | 12.5200 ms      | 2.8592 ms     | -77.168%   |
| code_action               | 11.0798 ms      | 1.9683 ms     | -82.233%   |

## Checklist

- [x] I have linked to any relevant issues.
- [x] I have commented my code, particularly in hard-to-understand areas.
- [ ] I have updated the documentation where relevant (API docs, the reference, and the Sway book).
- [ ] I have added tests that prove my fix is effective or that my feature works.
- [ ] I have added (or requested a maintainer to add) the necessary `Breaking*` or `New Feature` labels where relevant.
- [x] I have done my best to ensure that my PR adheres to [the Fuel Labs Code Review Standards](https://github.com/FuelLabs/rfcs/blob/master/text/code-standards/external-contributors.md).
- [x] I have requested a review from the relevant team or maintainers.